refactor: move helper functions out of the route module into mcp service file
